fix user files

+++ app/src/main/java/com/cretihoy/wordminded2/data/db/UserEntity.kt
@@ -14,5 +14,5 @@ data class UserEntity(
     val name: String,
 
     @ColumnInfo
-    val score: Long
+    val score: Int
 )
